refactor(media): previewing files components
